refactor: modularize webui-cli architecture
Restructure CLI into a maintainable multi-module layout:
- commands/mod.rs: Commands enum for easy subcommand addition
- commands/build.rs: Build args, logic, error hints, and tests
- output.rs: Shared Printer struct for styled terminal output
- main.rs: Thin entry point with dispatch only

Each command now owns its own error handling and contextual hints
via an execute()/run() pattern, keeping main.rs clean.
