+++ crates/webui-cli/src/commands/mod.rs
@@ -1,0 +1,9 @@
+pub mod build;
+
+use clap::Subcommand;
+
+#[derive(Subcommand)]
+pub enum Commands {
+    /// Build a WebUI application from an app folder
+    Build(build::BuildArgs),
+}
